# go-client-generated/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is API for cloud-based microservice"
  version: "1.0.0"
  title: "Simple Inventory API"
  contact:
    email: "newmail@mail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "virtserver.swaggerhub.com"
basePath: "/None368/Acronis/1.0.0"
...
schemes:
- "https"
paths:
...
  /logs:
    post:
      tags:
      - "Logs"
      summary: "Post logs"
      description: "Post logs to server"
      operationId: "PostLogs"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "logs"
        required: false
        schema:
          type: "string"
          format: "binary"
        x-exportParamName: "Logs"
      responses:
        200:
          description: "OK"
    delete:
      tags:
      - "Logs"
      summary: "Delete all logs"
      operationId: "deleteAllLogs"
      parameters: []
      responses:
        200:
          description: "OK"
  /logs/{id}:
    get:
      tags:
      - "Logs"
      summary: "Get tail of log"
      description: "Get logs by id from server"
      operationId: "GetLogsById"
      parameters:
      - name: "id"
        in: "path"
        description: "Logs id"
        required: true
        type: "integer"
        x-exportParamName: "Id"
      responses:
        200:
          description: "OK"
        404:
          description: "NOT FOUND"
    delete:
      tags:
      - "Logs"
      summary: "Delete logs by id"
      description: "Delete logs by id from server"
      operationId: "DeleteLogsById"
      parameters:
      - name: "id"
        in: "path"
        description: "Logs id"
        required: true
        type: "integer"
        x-exportParamName: "Id"
      responses:
        200:
          description: "OK"
        404:
          description: "NOT FOUND"
...